# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Koshagar cloud storage API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type FileItemType = typeof FileItemType[keyof typeof FileItemType];


export const FileItemType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface FileItem {
  id: number;
  name: string;
  type: FileItemType;
  mimeType: string;
  size: number;
  starred: boolean;
  trashed: boolean;
  /** @nullable */
  folderId?: number | null;
  ownerId: number;
  /** @nullable */
  shareToken?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  trashedAt?: string | null;
}

export interface FileListResponse {
  folders: FileItem[];
  files: FileItem[];
}

export interface FileUpdate {
  name?: string;
}

export interface UploadInput {
  name: string;
  mimeType: string;
  size: number;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  dataUrl?: string | null;
}

export interface StarInput {
  starred: boolean;
}

export interface TrashInput {
  trashed: boolean;
}

export interface MoveInput {
  /** @nullable */
  folderId?: number | null;
}

export interface FolderInput {
  name: string;
  /** @nullable */
  folderId?: number | null;
}

export interface BreadcrumbItem {
  id: number;
  name: string;
}

export interface FolderDetail {
  folder: FileItem;
  breadcrumb: BreadcrumbItem[];
}

export interface ShareInput {
  fileId: number;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  allowDownload?: boolean;
}

export interface ShareLink {
  id: number;
  token: string;
  fileId: number;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  allowDownload: boolean;
  viewCount?: number;
  downloadCount?: number;
  createdAt: string;
}

export interface PublicShareView {
  file: FileItem;
  allowDownload: boolean;
  sharedBy: string;
  /** @nullable */
  expiresAt?: string | null;
  requiresPassword?: boolean;
}

export interface ShareStats {
  viewCount: number;
  downloadCount: number;
}

export type ActivityEntryAction = typeof ActivityEntryAction[keyof typeof ActivityEntryAction];


export const ActivityEntryAction = {
  upload: 'upload',
  download: 'download',
  share: 'share',
  delete: 'delete',
  rename: 'rename',
  move: 'move',
  trash: 'trash',
  restore: 'restore',
  star: 'star',
} as const;

export interface ActivityEntry {
  id: number;
  action: ActivityEntryAction;
  fileName: string;
  /** @nullable */
  fileId?: number | null;
  userId: number;
  userName: string;
  createdAt: string;
}

export interface ActivityPage {
  entries: ActivityEntry[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}

export interface StorageUsage {
  usedBytes: number;
  totalBytes: number;
  fileCount: number;
}

export interface AdminStats {
  totalUsers: number;
  totalFiles: number;
  totalStorageBytes: number;
  activeToday: number;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  role: string;
  fileCount: number;
  storageBytes: number;
  createdAt: string;
  /** @nullable */
  lastActiveAt?: string | null;
}

export type ListFilesParams = {
/**
 * Folder ID (null for root)
 * @nullable
 */
folderId?: number | null;
/**
 * @nullable
 */
starred?: boolean | null;
/**
 * @nullable
 */
trash?: boolean | null;
/**
 * @nullable
 */
search?: string | null;
sort?: ListFilesSort;
order?: ListFilesOrder;
};

export type ListFilesSort = typeof ListFilesSort[keyof typeof ListFilesSort];


export const ListFilesSort = {
  name: 'name',
  createdAt: 'createdAt',
  size: 'size',
  updatedAt: 'updatedAt',
} as const;

export type ListFilesOrder = typeof ListFilesOrder[keyof typeof ListFilesOrder];


export const ListFilesOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetActivityParams = {
/**
 * @nullable
 */
cursor?: number | null;
limit?: number;
};

export type GetAdminActivityParams = {
limit?: number;
};