# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Koshagar cloud storage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a folder
      parameters:
        - name: folderId
          in: query
          schema:
            type: ["integer", "null"]
          description: Folder ID (null for root)
        - name: starred
          in: query
          schema:
            type: ["boolean", "null"]
        - name: trash
          in: query
          schema:
            type: ["boolean", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, createdAt, size, updatedAt]
            default: name
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: List of files and folders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Uploaded file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
...
  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
        "404":
          description: Not found
    patch:
      operationId: updateFile
      tags: [files]
      summary: Rename or update a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Permanently delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /files/{id}/star:
    patch:
      operationId: starFile
      tags: [files]
      summary: Toggle star on a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StarInput"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
  /files/{id}/trash:
    patch:
      operationId: trashFile
      tags: [files]
      summary: Move a file to trash or restore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrashInput"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
  /files/{id}/move:
    patch:
      operationId: moveFile
      tags: [files]
      summary: Move a file to a different folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveInput"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
  /folders:
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
  /folders/{id}:
    get:
      operationId: getFolder
      tags: [folders]
      summary: Get folder with breadcrumb path
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Folder with breadcrumb
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetail"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    FileItem:
      type: object
      required: [id, name, type, mimeType, size, starred, trashed, createdAt, updatedAt, ownerId]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [file, folder]
        mimeType:
          type: string
        size:
          type: integer
        starred:
          type: boolean
        trashed:
          type: boolean
        folderId:
          type: ["integer", "null"]
        ownerId:
          type: integer
        shareToken:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        trashedAt:
          type: ["string", "null"]
    FileListResponse:
      type: object
      required: [folders, files]
      properties:
        folders:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
    FileUpdate:
      type: object
      properties:
        name:
          type: string
    UploadInput:
      type: object
      required: [name, mimeType, size]
      properties:
        name:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        folderId:
          type: ["integer", "null"]
        dataUrl:
          type: ["string", "null"]
    StarInput:
      type: object
      required: [starred]
      properties:
        starred:
          type: boolean
    TrashInput:
      type: object
      required: [trashed]
      properties:
        trashed:
          type: boolean
    MoveInput:
      type: object
      properties:
        folderId:
          type: ["integer", "null"]
    FolderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        folderId:
          type: ["integer", "null"]
    FolderDetail:
      type: object
      required: [folder, breadcrumb]
      properties:
        folder:
          $ref: "#/components/schemas/FileItem"
        breadcrumb:
          type: array
          items:
            $ref: "#/components/schemas/BreadcrumbItem"
    BreadcrumbItem:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...