# lib/api-zod/src/generated/types/breadcrumbItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Koshagar cloud storage API
 * OpenAPI spec version: 0.1.0
 */

export interface BreadcrumbItem {
  id: number;
  name: string;
}