# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Koshagar cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItemType } from './fileItemType';

export interface FileItem {
  id: number;
  name: string;
  type: FileItemType;
  mimeType: string;
  size: number;
  starred: boolean;
  trashed: boolean;
  /** @nullable */
  folderId?: number | null;
  ownerId: number;
  /** @nullable */
  shareToken?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  trashedAt?: string | null;
}