# lib/api-zod/src/generated/types/fileListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Koshagar cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItem } from './fileItem';

export interface FileListResponse {
  folders: FileItem[];
  files: FileItem[];
}